Moving database stuff back to src/db
