+++ src/main/index.js
@@ -1,8 +1,6 @@
 'use strict'
 
 import { app, BrowserWindow } from 'electron'
-// import SurvivorsDatabase from '../renderer/survivors'
-// import SettlementsDatabase from '../renderer/settlements'
 
 /**
  * Set `__static` path to static files in production
@@ -32,20 +30,6 @@ function createWindow () {
   mainWindow.on('closed', () => {
     mainWindow = null
   })
-
-  // TEST CODE!
-  // console.log(app.getPath('userData'))
-
-  // console.log(smts)
-  // var survs = new SurvivorsDatabase(app.getPath('userData'), smts)
-
-  // smts.createNew((smt) => {
-  //   var smtID = smt._id
-  //   survs.addBase(smtID, { name: 'Alex', sex: 'M' }, (s) => {
-  //     console.log(s)
-  //     survs.getAll(smtID, (survs) => console.log(survs))
-  //   })
-  // })
 }
 
 app.on('ready', createWindow)
